refactoring: tried to decuple the state from the ui
