+++ app/src/main/java/org/stypox/tridenta/widget/MyAppWidget.kt
@@ -20,89 +20,34 @@ import androidx.glance.appwidget.state.updateAppWidgetState
 import androidx.glance.currentState
 import androidx.glance.preview.ExperimentalGlancePreviewApi
 import androidx.glance.preview.Preview
-import androidx.glance.state.PreferencesGlanceStateDefinition
-import dagger.hilt.android.EntryPointAccessors
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.getAndUpdate
 import kotlinx.coroutines.flow.update
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import org.stypox.tridenta.db.HistoryDao
 import org.stypox.tridenta.db.data.DbLine
 import org.stypox.tridenta.db.data.DbStop
 import org.stypox.tridenta.enums.Area
 import org.stypox.tridenta.enums.CardinalPoint
 import org.stypox.tridenta.enums.Direction
 import org.stypox.tridenta.enums.StopLineType
-import org.stypox.tridenta.extractor.ROME_ZONE_ID
-import org.stypox.tridenta.log.logError
 import org.stypox.tridenta.log.logInfo
-import org.stypox.tridenta.repo.LineTripsRepository
-import org.stypox.tridenta.repo.LinesRepository
 import org.stypox.tridenta.repo.data.UiStopTime
 import org.stypox.tridenta.repo.data.UiTrip
 import org.stypox.tridenta.ui.MainActivity
-import org.stypox.tridenta.ui.line_trips.LineTripsUiState
 import org.stypox.tridenta.widget.actions.NextTripAction
 import org.stypox.tridenta.widget.actions.PrevTripAction
 import org.stypox.tridenta.widget.actions.ReloadTripAction
 import org.stypox.tridenta.widget.actions.ToggleDirectionAction
-import org.stypox.tridenta.widget.actions.WidgetEntryPoint
 import org.stypox.tridenta.widget.actions.WidgetKeys
 import org.stypox.tridenta.widget.ui.LineTripsWidgetScreen
 import org.stypox.tridenta.widget.ui.TripViewGlance
 import java.time.OffsetDateTime
 import java.time.ZoneOffset
-import java.time.ZonedDateTime
-import kotlin.properties.Delegates
 
 class MyAppWidget : GlanceAppWidget() {
-    override val stateDefinition = PreferencesGlanceStateDefinition
-    private lateinit var tripsRepository: LineTripsRepository
-    private lateinit var linesRepository: LinesRepository
-    private lateinit var historyDao: HistoryDao
-
-    //private lateinit var referenceDateTime: ZonedDateTime
-    private val mutableUiState = MutableStateFlow(
-        LineTripsUiState(
-            line = null,
-            tripsInDayCount = 0,
-            tripIndex = 0,
-            trip = null,
-            prevEnabled = false,
-            nextEnabled = false,
-            referenceDateTime = ZonedDateTime.now().withZoneSameInstant(ROME_ZONE_ID),
-            directionFilter = Direction.ForwardAndBackward,
-            stopIdToHighlight = null,
-            stopTypeToHighlight = null,
-            loading = true,
-            error = false,
-        )
-    )
-    val uiState = mutableUiState.asStateFlow()
-    private var lineId by Delegates.notNull<Int>()
-    private lateinit var lineType: StopLineType
-
-    private var tripReloadJob: Job? = null
-
+    lateinit var model: WidgetModel
     override suspend fun provideGlance(
         context: Context, id: GlanceId
     ) {
-        val hiltEntryPoint =
-            EntryPointAccessors.fromApplication(context, WidgetEntryPoint::class.java)
-        tripsRepository = hiltEntryPoint.lineTripsRepository()
-        linesRepository = hiltEntryPoint.lineRepository()
-        historyDao = hiltEntryPoint.historyDao()
-
-        mutableUiState.update {
-            it.copy(
-                referenceDateTime = ZonedDateTime.now().withZoneSameInstant(ROME_ZONE_ID)
-            )
-        }
+        model = WidgetModel(context, id)
+        model.initState()
 
         val appWidgetId = GlanceAppWidgetManager(context).getAppWidgetId(id)
         val configIntent = Intent(context, WidgetConfigurationActivity::class.java).apply {
@@ -110,82 +55,67 @@ class MyAppWidget : GlanceAppWidget() {
             flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_MULTIPLE_TASK
         }
         provideContent {
-            val lineTripsUiState by uiState.collectAsState()
+            val lineTripsUiState by model.uiState.collectAsState()
             val prefs = currentState<Preferences>()
-            //var lineType by remember { mutableStateOf(StopLineType.Urban) }
 
-            lineId = prefs[WidgetKeys.LINE_ID] ?: -1
+            val lineId = prefs[WidgetKeys.LINE_ID] ?: -1
             val lineTypeString = prefs[WidgetKeys.LINE_TYPE]
-            if (lineTypeString != null) lineType = StopLineType.valueOf(lineTypeString)
             val tripIndex = prefs[WidgetKeys.TRIP_INDEX]
             val toggledDirection = prefs[WidgetKeys.TOGGLED_DIRECTION] ?: false
             val prevTripIndex = prefs[WidgetKeys.PREV_TRIP_INDEX]
             val refreshTimestamp = prefs[WidgetKeys.REFRESH_TIMESTAMP] ?: 0L
             val storedDirectionFilter = prefs[WidgetKeys.DIRECTION_FILTER]
-            val prevEnabledStored = prefs[WidgetKeys.PREV_ENABLED]
-            if (prevEnabledStored != null) mutableUiState.update {
-                it.copy(prevEnabled = prevEnabledStored)
-            }
-            val nextEnabledStored = prefs[WidgetKeys.NEXT_ENABLED]
-            if (nextEnabledStored != null) mutableUiState.update {
-                it.copy(
-                    nextEnabled = nextEnabledStored
-                )
-            }
             val tripsInDayCount = prefs[WidgetKeys.TRIPS_IN_DAY_COUNT]
-            if (tripsInDayCount != null) {
-                mutableUiState.update {
-                    it.copy(tripsInDayCount = tripsInDayCount)
-                }
-            }
 
-            LaunchedEffect(lineId, lineTypeString) { // retrieving line
+            LaunchedEffect(lineId) { // retrieving line
                 if (lineId == -1 || lineTypeString == null) {
                     return@LaunchedEffect
                 }
+                model.initState()
                 if (lineTripsUiState.line == null || lineTripsUiState.line!!.lineId != lineId) {
-                    loadLine(lineId, lineType)
+                    model.loadLine()
                 }
             }
 
             LaunchedEffect(
-                lineId, tripIndex, prevTripIndex, tripsInDayCount
+                model.lineId, tripIndex, prevTripIndex, tripsInDayCount
             ) { // retrieving trip
                 if (tripIndex == null) {
-                    cancelTripReloadJobAndLaunch {
-                        setReferenceDateTimeAsync(lineTripsUiState.referenceDateTime, context, id)
+                    model.cancelTripReloadJobAndLaunch {
+                        model.setReferenceDateTimeAsync(lineTripsUiState.referenceDateTime)
                     }
                     return@LaunchedEffect
                 }
-                loadIndex(tripIndex, context, id)
+                model.loadIndex(tripIndex)
+                model.initState()
             }
 
             LaunchedEffect(storedDirectionFilter) {
                 if (storedDirectionFilter == null || !toggledDirection) return@LaunchedEffect
                 val newDirectionFilter = Direction.valueOf(storedDirectionFilter)
-                mutableUiState.update { it.copy(directionFilter = newDirectionFilter) }
+                model.mutableUiState.update { it.copy(directionFilter = newDirectionFilter) }
 
                 if (newDirectionFilter == Direction.ForwardAndBackward) {
-                    val state = uiState.value
+                    val state = model.uiState.value
                     if (state.trip == null) {
                         // the trip can be null if there is no trip in that direction
-                        loadIndex(state.tripIndex, context, id)
+                        model.loadIndex(state.tripIndex)
                     } else {
                         // no need to load the trip, as it's already loaded
-                        mutableUiState.update {
+                        model.mutableUiState.update {
                             it.copy(
                                 prevEnabled = state.tripIndex > 0,
-                                nextEnabled = state.tripIndex < uiState.value.tripsInDayCount - 1,
+                                nextEnabled = state.tripIndex < model.uiState.value.tripsInDayCount - 1,
                                 directionFilter = newDirectionFilter,
                             )
                         }
                     }
 
                 } else {
-                    val state = uiState.value
+                    val state = model.uiState.value
                     if (state.trip?.direction != newDirectionFilter) {
                         // we need to load another trip, since the current one has the wrong direction
-                        loadIndex(state.tripIndex, context, id)
+                        model.loadIndex(state.tripIndex)
                         logInfo("state.tripIndex: ${state.tripIndex}")
                         logInfo("lineTripsUiState.tripIndex: ${lineTripsUiState.tripIndex}")
                         updateAppWidgetState(context, id) { prefs ->
@@ -193,12 +123,13 @@ class MyAppWidget : GlanceAppWidget() {
                             prefs[WidgetKeys.PREV_TRIP_INDEX] = state.tripIndex
                             prefs[WidgetKeys.TRIPS_IN_DAY_COUNT] = lineTripsUiState.tripsInDayCount
                             prefs[WidgetKeys.IS_INITIAL_DATA_LOADED] = true
-                            prefs[WidgetKeys.DIRECTION_FILTER] = lineTripsUiState.directionFilter.name
+                            prefs[WidgetKeys.DIRECTION_FILTER] =
+                                lineTripsUiState.directionFilter.name
                             prefs[WidgetKeys.PREV_ENABLED] = lineTripsUiState.tripIndex > 0
                             prefs[WidgetKeys.NEXT_ENABLED] =
                                 lineTripsUiState.tripIndex < lineTripsUiState.tripsInDayCount - 1
                         }
-                        this@MyAppWidget.update(context,id)
+                        this@MyAppWidget.update(context, id)
                     }
                 }
                 updateAppWidgetState(context, id) { prefs ->
@@ -206,10 +137,18 @@ class MyAppWidget : GlanceAppWidget() {
                 }
             }
 
+            if (storedDirectionFilter != null && Direction.valueOf(storedDirectionFilter) != lineTripsUiState.directionFilter) {
+                model.mutableUiState.update {
+                    it.copy(
+                        directionFilter = Direction.valueOf(storedDirectionFilter)
+                    )
+                }
+            }
+
             LaunchedEffect(refreshTimestamp) { // updating trip
                 if (refreshTimestamp == 0L || tripIndex == null || lineTripsUiState.trip == null) return@LaunchedEffect
                 //updateTrip(tripIndex, mutableUiState, lineTripsUiState.trip!!)
-                cancelTripReloadJobAndLaunch { onReloadAsync(context, id) }
+                model.cancelTripReloadJobAndLaunch { model.onReloadAsync() }
             }
 
             logInfo("${System.currentTimeMillis()}")
@@ -240,275 +179,6 @@ class MyAppWidget : GlanceAppWidget() {
                     isFavorite = lineTripsUiState.line?.isFavorite ?: false
                 )
             }
-        }
-    }
-
-    private fun loadLine(lineId: Int, lineType: StopLineType) {
-        // this job is independent from trip reloading jobs, so don't use cancelReloadJobAndLaunch
-        CoroutineScope(Dispatchers.IO).launch {
-            val line = withContext(Dispatchers.IO) {
-                try {
-                    linesRepository.getUiLine(lineId, lineType).also {
-                        if (it == null) {
-                            logError(
-                                "UI line (${lineId}, ${lineType}) not found"
-                            )
-                        }
-
-                        // register a view for this line (assuming loadLine is called once)
-                        historyDao.registerAccessed(true, lineId, lineType)
-                    }
-                } catch (e: Throwable) {
-                    logError("Could not load UI line (${lineId}, ${lineType})", e)
-                    null
-                }
-            }
-            mutableUiState.update { it.copy(line = line) }
-        }
-    }
-
-    private fun cancelTripReloadJobAndLaunch(tripLoadingFunction: suspend () -> Unit) {
-        tripReloadJob?.cancel()
-        tripReloadJob = CoroutineScope(Dispatchers.IO).launch {
-            // clear errors and set the state to "loading" before starting to load
-            mutableUiState.update { it.copy(loading = true, error = false) }
-            tripLoadingFunction()
-            // set "loading" to false when loading finishes (errors are set inside the function)
-            mutableUiState.update { it.copy(loading = false) }
-        }
-    }
-
-    private fun loadIndex(index: Int, context: Context, id: GlanceId) {
-        logInfo("loadIndex: $index")
-        if (index >= 0 && index < uiState.value.tripsInDayCount) {
-            // only cancel any currently running job if there is something to do; the above
-            // condition will be false e.g. when there are no trips in a day (but not only for that)
-            cancelTripReloadJobAndLaunch {
-                loadIndexAsync(index, context, id)
-            }
-        }
-    }
-
-    private suspend fun loadIndexAsync(index: Int, context: Context, id: GlanceId) {
-        // hide the current trip, as it's going to change
-        val prevState = mutableUiState.getAndUpdate {
-            it.copy(
-                tripIndex = index,
-                trip = null,
-                prevEnabled = index > 0,
-                nextEnabled = index < uiState.value.tripsInDayCount - 1,
-            )
-        }
-
-        // load the trip, and show it right after it gets loaded (see the related functions)
-        val (trip, network) = if (prevState.directionFilter == Direction.ForwardAndBackward) {
-            loadIndexNoFilterAsync(index, context, id)
-        } else {
-            loadIndexDirectionAsync(index, prevState, context, id)
-        }
-
-        if (!network && trip != null && trip.completedStops < trip.stopTimes.size) {
-            // after showing the (possibly) outdated trip fast, reload it to show latest updates
-            // (but reload it only if there actually is a trip and it is not completed)
-            onReloadAsync(context, id)
-        }
-    }
-
-    private suspend fun setReferenceDateTimeAsync(
-        referenceDateTimeCurrentZone: ZonedDateTime, context: Context, id: GlanceId
-    ) {
-        val referenceDateTime = referenceDateTimeCurrentZone.withZoneSameInstant(ROME_ZONE_ID)
-        mutableUiState.update {
-            it.copy(
-                tripsInDayCount = 0,
-                tripIndex = 0,
-                trip = null,
-                prevEnabled = false,
-                nextEnabled = false,
-                referenceDateTime = referenceDateTime,
-            )
-        }
-
-        var error = false
-        val (tripsInDayCount, tripIndex, trip) = withContext(Dispatchers.IO) {
-            try {
-                tripsRepository.getUiTrip(
-                    lineId = lineId,
-                    lineType = lineType,
-                    referenceDateTime = referenceDateTime,
-                    directionFilter = uiState.value.directionFilter,
-                )
-            } catch (e: Throwable) {
-                logError(
-                    "Could not load trip for UI line (${mutableUiState.value.line?.lineId}, " + "${mutableUiState.value.line?.type}) at time $referenceDateTime",
-                    e
-                )
-                error = true
-                Triple(0, 0, null)
-            }
-        }
-
-        mutableUiState.update {
-            it.copy(
-                tripsInDayCount = tripsInDayCount,
-                tripIndex = tripIndex,
-                trip = trip,
-                prevEnabled = tripIndex > 0,
-                nextEnabled = tripIndex < tripsInDayCount - 1,
-                referenceDateTime = referenceDateTime,
-                error = error,
-            )
-        }
-        updateAppWidgetState(context, id) { prefs ->
-            prefs[WidgetKeys.TRIP_INDEX] = tripIndex
-            prefs[WidgetKeys.TRIPS_IN_DAY_COUNT] = tripsInDayCount
-            prefs[WidgetKeys.IS_INITIAL_DATA_LOADED] = true
-            prefs[WidgetKeys.PREV_ENABLED] = tripIndex > 0
-            prefs[WidgetKeys.NEXT_ENABLED] = tripIndex < tripsInDayCount - 1
-        }
-        logInfo("setReferenceDateTimeAsync tripIndex: $tripIndex")
-    }
-
-    private suspend fun onReloadAsync(context: Context, id: GlanceId) {
-        val previousTrip = uiState.value.trip
-        if (previousTrip == null) {
-            // this could happen if an error happened while loading initial/more trips
-            if (uiState.value.tripsInDayCount > 0) {
-                // more trips failed loading, try to load again the currently set index
-                loadIndexAsync(uiState.value.tripIndex, context, id)
-            } else {
-                // initial trips failed loading, try to load again the current day
-                setReferenceDateTimeAsync(uiState.value.referenceDateTime, context, id)
-            }
-            return
-        }
-
-        val trip = withContext(Dispatchers.IO) {
-            try {
-                tripsRepository.reloadUiTrip(
-                    uiTrip = previousTrip,
-                    index = uiState.value.tripIndex,
-                    referenceDateTime = uiState.value.referenceDateTime
-                )
-            } catch (e: Throwable) {
-                logError(
-                    "Could not load trip ${previousTrip.tripId} for UI line " + "(${mutableUiState.value.line?.lineId}, ${mutableUiState.value.line?.type})",
-                    e
-                )
-                null
-            }
-        }
-
-        if (trip == null) {
-            // keep previous trip intact, we don't want to hide information that we do have!
-            mutableUiState.update { it.copy(error = true) }
-        } else {
-            mutableUiState.update { it.copy(trip = trip) }
-        }
-    }
-
-    private suspend fun loadIndexNoFilterAsync(
-        index: Int, context: Context, id: GlanceId
-    ): Pair<UiTrip?, Boolean> {
-        val res = withContext(Dispatchers.IO) {
-            try {
-                tripsRepository.getUiTrip(
-                    lineId = lineId,
-                    lineType = lineType,
-                    referenceDateTime = uiState.value.referenceDateTime,
-                    index = index,
-                )
-            } catch (e: Throwable) {
-                logError(
-                    "Could not load trip at index $index for UI line " + "(${lineId}, ${lineType})",
-                    e
-                )
-                Pair(null, true /* <- useless when trip == null */)
-            }
-        }
-
-        // show the trip even if loadedFromNetwork is false (in which case it could be outdated)
-        mutableUiState.update {
-            it.copy(
-                trip = res.first,
-                error = res.first == null,
-            )
-        }
-        if (res.first != null) {
-            updateAppWidgetState(context, id) { prefs ->
-                prefs[WidgetKeys.TRIP_INDEX] = index
-                prefs[WidgetKeys.PREV_ENABLED] = index > 0
-                prefs[WidgetKeys.NEXT_ENABLED] = index < uiState.value.tripsInDayCount - 1
-            }
-        }
-
-        return res
-    }
-
-
-    private suspend fun loadIndexDirectionAsync(
-        index: Int, prevState: LineTripsUiState, context: Context, id: GlanceId
-    ): Pair<UiTrip?, Boolean> {
-        val res = withContext(Dispatchers.IO) {
-            try {
-                tripsRepository.getUiTripWithDirection(
-                    lineId = lineId,
-                    lineType = lineType,
-                    referenceDateTime = uiState.value.referenceDateTime,
-                    index = index,
-                    direction = prevState.directionFilter,
-                    prevIndex = prevState.tripIndex,
-                )
-            } catch (e: Throwable) {
-                logError(
-                    "Could not load trip at index $index for UI line " + "(${mutableUiState.value.line?.lineId}, ${mutableUiState.value.line?.type})" + "in direction ${prevState.directionFilter}",
-                    e
-                )
-                null
-            }
-        }
-
-        if (res == null) {
-            // no trip could be loaded in the set direction, restore previous state,
-            // but update prevEnabled or nextEnabled
-            mutableUiState.update {
-                it.copy(
-                    tripIndex = prevState.tripIndex,
-                    trip = prevState.trip,
-                    prevEnabled = if (index < prevState.tripIndex) false else prevState.prevEnabled,
-                    nextEnabled = if (index > prevState.tripIndex) false else prevState.nextEnabled,
-                )
-            }
-            updateAppWidgetState(context, id) { prefs ->
-                prefs[WidgetKeys.TRIP_INDEX] = prevState.tripIndex
-                if (index < prevState.tripIndex) {
-                    prefs[WidgetKeys.PREV_ENABLED] = false
-                }
-                if (index > prevState.tripIndex) {
-                    prefs[WidgetKeys.NEXT_ENABLED] = false
-                }
-            }
-
-            return Pair(null, true /* <- useless when trip == null */)
-
-        } else {
-            val (trip, newIndex, loadedFromNetwork) = res
-            mutableUiState.update {
-                it.copy(
-                    tripIndex = newIndex,
-                    trip = trip,
-                    prevEnabled = newIndex > 0,
-                    nextEnabled = newIndex < uiState.value.tripsInDayCount - 1,
-                )
-            }
-            updateAppWidgetState(context, id) { prefs ->
-                prefs[WidgetKeys.TRIP_INDEX] = newIndex
-                prefs[WidgetKeys.PREV_TRIP_INDEX] = prevState.tripIndex
-                prefs[WidgetKeys.PREV_ENABLED] = newIndex > 0
-                prefs[WidgetKeys.NEXT_ENABLED] = newIndex < uiState.value.tripsInDayCount - 1
-            }
-
-            return Pair(trip, loadedFromNetwork)
         }
     }
 }
